dom/u2f/tests/test_bind.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test that bind() can be called on u2f.register/sign</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    SpecialPowers.pushPrefEnv({ "set": [["security.webauth.u2f", true]] },
                              doTest);
    function doTest() {
        // For some reason, exceptions from inside here do not cause a test
        // failure!  Run the actual code we want off something that will report
        // errors properly.
        SimpleTest.executeSoon(function() {
            is(typeof(u2f.register.bind(u2f)), "function",
               "Should be able to bind u2f.register");
            is(typeof(u2f.sign.bind(u2f)), "function",
               "Should be able to bind u2f.sign");
            SimpleTest.finish();
        });
    }
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
